+++ spec/filters/unsupported/bignum.rb
@@ -44,4 +44,9 @@ opal_filter "Bignum" do
   fails "Rational#** when passed Bignum returns Rational(-1) when self is Rational(-1) and the exponent is positive and odd"
   fails "Rational#** when passed Bignum returns Rational(1) when self is Rational(-1) and the exponent is positive and even"
   fails "Rational#round with a precision > 0 doesn't fail when rounding to an absurdly large positive precision"
+  fails "Marshal.load loads a Bignum 2**90"
+  fails "Marshal.load for a Integer loads an Integer 2361183241434822606847"
+  fails "Marshal.load for a Integer loads an Integer -2361183241434822606847"
+  fails "Marshal.dump with a Bignum dumps a Bignum"
+  fails "Marshal.dump with a Bignum dumps a Bignum"
 end